Converted ajs directive to ajs component
