+++ core/templates/components/state-editor/state-editor.directive.ts
@@ -29,7 +29,7 @@ require(
   'state-skill-editor.component.ts');
 require(
   'components/state-editor/state-responses-editor/' +
-  'state-responses.directive.ts');
+  'state-responses.component.ts');
 require(
   'components/state-editor/state-solution-editor/' +
   'state-solution-editor.directive.ts');
